@@ -3,30 +3,52 @@
   "version": "0.8.6",
   "license": "MIT",
   "scripts": {
-    "start": "webpack-dev-server --hot --inline",
-    "prebuild": "rimraf dist",
-    "build": "webpack",
-    "reinstall": "npm i rimraf && rimraf node_modules && npm uninstall -g elm && npm i -g elm && npm i && elm package install"
+    "webpack": "webpack",
+    "webpack-dev-server": "webpack-dev-server",
+    "prod": "npm run webpack -- --env.mode production",
+    "prod:analyze": "npm run prod -- --env.presets analyze",
+    "prod:compress": "npm run prod -- --env.presets compress",
+    "dev": "npm run webpack-dev-server -- --env.mode development",
+    "start": "npm run dev",
+    "build": "npm run prod",
+    "elm-analyse": "elm-analyse -s -o",
+    "preprod": "rimraf dist",
+    "lint": "stylelint src/assets/styles"
   },
   "devDependencies": {
-    "autoprefixer": "^6.7.7",
-    "bootstrap-sass": "^3.3.7",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.27.3",
-    "elm": "^0.18.0",
-    "elm-webpack-loader": "^4.3.0",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.10.1",
-    "html-webpack-plugin": "^2.28.0",
-    "jquery": "^3.2.1",
-    "node-sass": "^4.5.1",
-    "postcss-loader": "^1.3.3",
-    "rimraf": "^2.6.1",
-    "sass-loader": "^6.0.3",
-    "style-loader": "^0.16.0",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.3.1",
-    "webpack-dev-server": "^2.4.2",
-    "webpack-merge": "^4.1.0"
+    "autoprefixer": "^9.1.0",
+    "compression-webpack-plugin": "^1.1.12",
+    "copy-webpack-plugin": "^4.5.2",
+    "css-loader": "^1.0.0",
+    "elm-analyse": "^0.15.0",
+    "elm-hot-webpack-loader": "^1.0.1",
+    "elm-webpack-loader": "^5.0.0",
+    "file-loader": "^2.0.0",
+    "html-webpack-plugin": "^3.0.6",
+    "imagemin-webpack-plugin": "^2.1.5",
+    "mini-css-extract-plugin": "^0.4.2",
+    "node-sass": "^4.9.3",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "popper.js": "^1.14.4",
+    "postcss-loader": "^3.0.0",
+    "purify-css": "^1.2.5",
+    "purifycss-webpack": "^0.7.0",
+    "rimraf": "^2.6.2",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.0",
+    "stylelint": "^9.5.0",
+    "stylelint-config-standard": "^18.2.0",
+    "stylelint-webpack-plugin": "^0.10.4",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "webpack": "^4.17.1",
+    "webpack-bundle-analyzer": "^2.11.1",
+    "webpack-cli": "^3.1.0",
+    "webpack-dashboard": "^2.0.0",
+    "webpack-dev-server": "^3.1.7",
+    "webpack-merge": "^4.1.4"
+  },
+  "dependencies": {
+    "id3-parser": "^2.0.0",
+    "socket.io-client": "^2.1.1"
   }
 }
