@@ -1,127 +1,46 @@
-var path = require('path');
-var webpack = require('webpack');
-var merge = require('webpack-merge');
-var HtmlWebpackPlugin = require('html-webpack-plugin');
-var autoprefixer = require('autoprefixer');
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var CopyWebpackPlugin = require('copy-webpack-plugin');
+const webpack = require('webpack');
+const webpackMerge = require('webpack-merge');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const CopyWebpackPlugin = require('copy-webpack-plugin')
+const StyleLintPlugin = require('stylelint-webpack-plugin');
 
+const modeConfig = env => require(`./build-utils/webpack.${env}`)(env);
+const presetConfig = require("./build-utils/loadPresets");
 
-const prod = 'production';
-const dev = 'development';
+module.exports = ({ mode, presets } = { mode: "production", presets: [] }) => {
+  console.log(`Building for: ${mode}`);
 
-// determine build env
-const TARGET_ENV = process.env.npm_lifecycle_event === 'build' ? prod : dev;
-const isDev = TARGET_ENV == dev;
-const isProd = TARGET_ENV == prod;
+  return webpackMerge(
+    {
+      mode,
+        resolve: {
 
-// entry and output path/filename variables
-const entryPath = path.join(__dirname, 'src/static/index.js');
-const outputPath = path.join(__dirname, 'dist');
-const outputFilename = isProd ? '[name]-[hash].js' : '[name].js'
-
-console.log('WEBPACK GO! Building for ' + TARGET_ENV);
-
-// common webpack config (valid for dev and prod)
-var commonConfig = {
-    output: {
-        path: outputPath,
-        filename: `static/js/${outputFilename}`,
-    },
-    resolve: {
-        extensions: ['.js', '.elm'],
-        modules: ['node_modules']
-    },
-    module: {
+        },
+      module: {
         noParse: /\.elm$/,
-        rules: [{
+        rules: [
+          {
             test: /\.(eot|ttf|woff|woff2|svg)$/,
-            use: 'file-loader?publicPath=../../&name=static/css/[hash].[ext]'
-        }]
-    },
-    plugins: [
-        new webpack.LoaderOptionsPlugin({
-            options: {
-                postcss: [autoprefixer()]
-            }
-        }),
-        new HtmlWebpackPlugin({
-            template: 'src/static/index.html',
-            inject: 'body',
-            filename: 'index.html'
-        })
-    ]
-}
-
-// additional webpack settings for local env (when invoked by 'npm start')
-if (isDev === true) {
-    module.exports = merge(commonConfig, {
-        entry: [
-            'webpack-dev-server/client?http://localhost:8080',
-            entryPath
-        ],
-        devServer: {
-            // serve index.html in place of 404 responses
-            historyApiFallback: true,
-            contentBase: './src',
-            hot: true
-        },
-        module: {
-            rules: [{
-                test: /\.elm$/,
-                exclude: [/elm-stuff/, /node_modules/],
-                use: [{
-                    loader: 'elm-webpack-loader',
-                    options: {
-                        verbose: true,
-                        warn: true,
-                        debug: true
-                    }
-                }]
-            },{
-                test: /\.sc?ss$/,
-                use: ['style-loader', 'css-loader', 'postcss-loader', 'sass-loader']
-            }]
-        }
-    });
-}
-
-// additional webpack settings for prod env (when invoked via 'npm run build')
-if (isProd === true) {
-    module.exports = merge(commonConfig, {
-        entry: entryPath,
-        module: {
-            rules: [{
-                test: /\.elm$/,
-                exclude: [/elm-stuff/, /node_modules/],
-                use: 'elm-webpack-loader'
-            }, {
-                test: /\.sc?ss$/,
-                use: ExtractTextPlugin.extract({
-                    fallback: 'style-loader',
-                    use: ['css-loader', 'postcss-loader', 'sass-loader']
-                })
-            }]
-        },
-        plugins: [
-            new ExtractTextPlugin({
-                filename: 'static/css/[name]-[hash].css',
-                allChunks: true,
-            }),
-            // new CopyWebpackPlugin([{
-            //     from: 'src/static/img/',
-            //     to: 'static/img/'
-            // }]),
-
-            // extract CSS into a separate file
-            // minify & mangle JS/CSS
-            new webpack.optimize.UglifyJsPlugin({
-                minimize: true,
-                compressor: {
-                    warnings: false
-                }
-                // mangle:  true
-            })
+            use: 'file-loader?publicPath=../../&name=assets/css/[hash].[ext]',
+          },
         ]
-    });
-}
+      },
+
+      plugins: [
+        new HtmlWebpackPlugin({
+          template: 'src/assets/index.html',
+          inject: 'body',
+          filename: 'index.html',
+        }),
+
+
+
+        new StyleLintPlugin(),
+
+
+      ]
+    },
+    modeConfig(mode),
+    presetConfig({ mode, presets }),
+  )
+};
